Implement logic to find old E-needs-mcve issues to close
This is the first commit in a series of commits to implement the
automatic triaging of old E-needs-mcve issues that was proposed and
discussed in [t-release/triage] and cross-posted to [T-compiler].

This commit only implements the logic to find what the issues to close,
and prints that info to stdout for inspection. Think of it as a dry run.

After we have convinced ourselves this logic works as it should, we can
implement the final steps:
* Actually close the issue
* Report closes to "triagebot closed issues" topic in "t-release/triage" Zulip

[t-release/triage]: https://rust-lang.zulipchat.com/#narrow/stream/242269-t-release.2Ftriage/topic/auto-close.20E-needs-mcve/near/400273684
[t-compiler]: https://rust-lang.zulipchat.com/#narrow/stream/131828-t-compiler/topic/auto.20closing.20E-meeds-mcve/near/399663832